requirements.txt
requests==2.31.0
environs==9.5.0
python-telegram-bot==13.15
google-cloud-dialogflow==2.23.2
google-cloud-api-keys==0.5.3